direct: Drop mutex usage from query paths
This was adopted somewhat straight from the older vala code
implementing the direct connection, it seemed unnecessary around
the time for query paths, it does seem unnecessary now.

The query paths can happen unlocked since they do not need any
exclusive access to resources, nor any synchronization with other
queries. Especially, the serialization paths may take long to
complete, so it does not make any sense to hold the mutex for so
long in these places.

And to avoid this mistake from happening again, rename the mutex
to make it clear it only applies to update paths.