libtracker-sparql: Avoid "unused return value" warnings from g_string_free()
With recent glib, g_string_free(..., FALSE) expects the return value containing
the unfreed string to be used. This did not really play along in the places we
combine it with sqlite3_result_text(), since we need to give it both string and
length, so we used the GString as a hollow container (i.e. freeing afterwards).

Change the pattern somewhat, so we use the return value from g_string_free()
as glib does expect of us.