libtracker-sparql: Port DBus endpoint to service/graph filter API
Replace the internal API to make use of the new TrackerEndpoint
access control capabilities and rewriting queries to include the expected
constraint clauses.

The portal has also been ported to the new API, filtering access to all
services and applying the graph access restrictions as specified by the
sandboxed application through it.