libtracker-sparql: Decompose TrackerResource in bus batches
Instead of sending it as a raw set of SPARQL updates, that do not
allow any form of caching or optimizations, decompose TrackerResource
into:
- The individual delete operations for properties being rewritten,
  as statements.
- The RDF data in Trig format.

This will allow the caching of the individual simple statements to
delete properties on the other side, with a high usability rate,
and the RDF will be handled through deserialization. Dropping the
frequency where SPARQL parsing is required to a minimum.