libtracker-sparql: Decouple FTS initialization from ontologies
Since we just need the FTS properties for fts:offsets being called
at runtime, we can pass the TrackerDataManager (that we know early
in startup) and let it figure out the FTS properties from there
when it's actually needed.

This is just the refactor to make that possible, there's no
functional changes in this commit.