libtracker-sparql: Avoid regression in renaming FTS5 tables
SQLite 3.44.x has a regression with the renaming of FTS5 tables
within a transaction (https://sqlite.org/forum/forumpost?udc=1&name=4dbfcd293d).

While this change has a workaround smell, we actually do not need the
tables to be renamed, since we drop+create the FTS5 tables if ontology
changes affect FTS fields. The original table is already ensured to be
dropped when we "alter" the FTS table. Avoiding this rename will work
for all SQLite versions.