libtracker-sparql: Refactor FTS table creation calls
Since we used to initialize FTS and create tables in one shot, the
function involved had a "create" argument to specify whether tables
had to be created, or they were there already.

Since FTS initialization moved elsewhere, this argument in this function
is just an early return. We can avoid the calls in the first place where
that argument is FALSE (i.e. regular initialization paths on an already
existing database).

This is a refactor with no functional changes.