libtracker-sparql: Improve/fix blank node handling of Turtle deserializer
Currently, we give every blank node label an urn:uuid: fake IRI that
gets promoted into a "real" IRI instead of a blank node. This is not
up to spec, nor our TrackerBatch guarantees that the blank node label
mapping is global to all of it, as this mapping is local to a deserializer.

To fix this, pass blank node labels as-is, and identify them with the
correct TrackerSparqlValueType for recurrences both in subject and object
roles. We now only generate unique blank node labels for anonymous blank
nodes (e.g. [] in Turtle) and ensure these start with "_:" to preserve
their role as blank nodes.

We also now rely on the global blank node map, passed as an argument
during deserialization, like it happens with other batch operations.