libtracker-sparql: Forward D-Bus UpdateArray FD as a TrackerBatch
Add/use a direct path for a D-Bus UpdateArray FD to be processed
directly by the core machinery, instead of converted to a TrackerBatch
by the TrackerEndpoint, then committed.

This minimizes the amount of memory necessary to hold this batch
operation, since it can be processed in small chunks from the FD
directly, instead of piled up in RAM before flushing.