libtracker-sparql: Preserve blank node labels as-is serializing Ttl/Trig
Right now, we convert the "_:" prefix to "__" and append our own "_:" to make
them a blank label. This results in possible blank node label mismatches if this
Ttl/Trig is run with other operations in a batch.

Preserve the blank node labels as-is, so that the blank node labels in the
same batch see the the same blank nodes locally.