cli: Split CLI commands into separate binaries
The help text is now provided via private `.desktop` files, this allows
the existing translations to be reused while also allowing
tracker-miners to provide help text for its own CLI subcommands.

This avoids the issues with double linking that cause crashes, e.g.
<https://gitlab.gnome.org/GNOME/tracker/-/issues/420>.

As part of this, make source files and deps explicit per CLI command.
The main goal here is to highlight cases where commands link against
the private libtracker-sparql library, as this is a kludge and causes
major problems with GObject state if it ends up loaded into the
process twice via dlopen().