core: Keep TrackerDataManager in "triples" virtual table implementation
We will be shuffling the creation of the TrackerOntologies to a later point,
make this code account for this by keeping the TrackerDataManager and getting
the TrackerOntologies when needed. This is practically unused at database
initialization time, so the TrackerOntologies will be guaranteed to be there
when this virtual table is used.