core: Perform database updates over TrackerOntologyChange
Rewrite the code performing database creation/changes so it
happens via the following steps:
- A TrackerOntologies is created from the RDF ontology given on
  TrackerSparqlConnection creation.
- If the database already exists, a TrackerOntologies is introspected
  from the database.
- The two ontologies are diff'ed, and a series of TrackerOntologyChange
  is created from the ontology differences.
- The TrackerDataManager applies each of these standalone changes to
  the database.

This results in simpler code and database changes that are
easier to follow, clearer error paths, and support for several
types of changes not previously supported.

Since our code now handles kinds of database changes that were
previously forbidden or mishandled, some of the ontology change
tests had to be adapted for these changes. Make the minimal adaptions
here to have the test suite pass, the additional guarantees will
be further exercised and documented in later commits.