core: Propagate an error on corner case during resource URI insertion
We are optimistic in the insertion of URIs and insert first, then look up
a pre-existing ID if the insertion failed due to UNIQUE constraints on
the Uri column.

However, insertion might fail due to other circumstances like lack of
disk space, and the later query for the ID might return 0 without errors.
We should handle this as an error situation, but no error is set since
we account for errors happening (and creating/destroying a GError on
such common paths would be taxing). Set an error in this situation so
that it's propagated and handled correctly, even though we lost the
original cause.