ci: Avoid duplicate pipelines
Instead of using the templates provided by gitlab, which have
limited configurability, adopt the approach taken at Mutter, where
duplicate push/merge-request pipelines are prevented, but there is
still a manual escape hatch for running a branch pipeline.

For this, the "repo-sanity" phase has moved to a different stage
prior to everything else, and will require manual triggering if
running on a branch that is not the default one, or one of our
release branches.