core: Handle the update buffer growing beyond the expected limits
Currently, the update machinery works on fairly small fixed buffers,
and is optimized for the buffer never growing beyond those limits.
Nominally, pointers to elements of a fixed size GArray are used as
keys in a hashtable, so any realloc in the array due to the buffer
overgrowing would render this hashtable data invalid.

Handle indefinite growth, and make this array of elements an array
of arrays of elements. This way we keep low sizes by default, allow
growing without invalidating previously existing pointers (by adding
new small sized buffer chunks), and during regular operation (i.e.
with the buffers not overgrowing) this just takes a couple more memory
allocations.