ci: Remove the 'ubuntu' user from Ubuntu container
Most containers do not have any precreated user account besides 'root'.
The Ubuntu image does and ships /etc/subgid and /etc/subuid to enable
this to work with rootless Podman. However, the way the ci-templates use
Podman means this can't work, and we see permission errors when trying
to upload the container.