docs: Wait for parser module before building ontology documentation
This is an indirect dependency (since we load the ontology to an
in-memory database to introspect it), and sometimes Meson gets ahead
of itself trying to build the documentation before these modules are
built.

Add these build_tgt dependencies right on the custom_target() call
to generate these docs, since we cannot make the common library nor
the ontology doc generator to wait on these.