cli: Reunite CLI tools again into a single executable
Make these tools a single executable, since we do no longer expect
this CLI to be extended by external projects.

At the same time, the "sql" subcommand has been outcast, as it is
merely a debugging tool. It is now installed independently in
libexecdir as "tinysparql-sql", but it is no longer a subcommand
of the main "tinysparql" command.