web-ide: Avoid serving web-ide on /sparql route
Previously in this merge request all the GET requests were served with either the web-ide or a 404 page.
But according to https://www.w3.org/TR/2013/REC-sparql11-service-description-20130321/ a request to /sparql without any query should be responded with a description of the sparql endpoint.
This change ensures that the server serves static files only on GET requests on routes other than /sparql.