core: Avoid flush prevention on INSERT OR REPLACE {} WHERE {}
This mechanism was added so that DELETE{}INSERT{}WHERE{} updates
did not change the resultset as it was being iterated, but it
missed that these paths are somewhat shared with our custom
INSERT OR REPLACE syntax (if used together with a WHERE clause)
this resulted in these update queries possibly missing some of
their updates.

Change these paths to explicitly avoid this INSERT OR REPLACE
syntax (handled through the TRACKER_UPDATE_UPDATE update type)
so that changes are flushed when necessary for the later queries
to find.

Fixes: f436db08b ("core: Avoid flushing updates within delete+insert+where queries")
Closes: https://gitlab.gnome.org/GNOME/gnome-music/-/issues/618