core: Allow ontology files in all our supported RDF formats
So far, we have just allowed the Turtle data format, and the ".ontology"
extension for ontology data. Extend this so we allow all our supported
RDF formats, provided that the files use the right extensions.

For backwards compatibility, the ".ontology" extension is still handled,
and it still results in using the Turtle deserializer.