libtracker-sparql: Avoid linking libtracker-common
There may indeed be unintended side effects doubly linking a static library in a module and the library that loads it.
Thus it is better to avoid linking libtracker-commmon with libtracker-sparql/remote.

This change removes this dependency and inlines the relevant code from libtracker-common/tracker-debug.c and libtracker-common/tracker-debug.h that we need for debugging HTTP requests.