core: Add a name for the (currently anonymous) default graph
The SPARQL/RDF specs doesn't like much when things are anonymous, this is
true for blank nodes, as well as graphs. Currently our implicit default graph
is completely anonymous, which means it appears in queries using the SPARQL
default graph pattern, but otherwise has no means to refer to it specifically.
This extends to FROM/FROM NAMED/WITH/USING/GRAPH/etc clauses that work on
graph IRIs.

The implicit default graph cannot be referenced anywhere there, so this limits
what we are able to do with it, e.g. we cannot query for data exclusively from
it, discarding other every graphs.

This doesn't sit well with the spec (although it never explicitly says that
all graphs should have name), and is inconvenient to us. This commit gives it
the "nrl:DefaultGraph" name, but does not hook it anywhere except to tiptoe over
the graph. Later commits will drop the special casing of the "main" SQLite
database, and make this closer to regular graphs.