core: Expose default graph to internals
Let the nrl:DefaultGraph be propagated through the set of known
named graphs of the TrackerDataManager, and use it to handle the
default unnamed graph in the places where it's used.

We still don't expose the default graph as a named one in the
upper layers, and as messy as things still are, this allows to
handle the main database in the same loop that handles all other
graphs.