core: Add a new internal function to get the graphs that apply
During SPARQL query processing, we keep the effective graphs that
apply when handling specific portions of a query. Add a new method
with a GraphSet argument (either NAMED or DEFAULT) that will figure
out the right graphs, after filtering due to policies, and the graph
sets specified by FROM/FROM NAMED if any.

This is at the moment unused, but will replace the current method
in later commits.