docs: Refer to nrl:DefaultGraph in docs
Change the "SPARQL and TinySPARQL" docs section about the default
graph, to refer to it and the adopted semantics to remain backwards
compatible.

In short: The nrl:DefaultGraph graph is to most effects anonymous as
it used to be, our default set for FROM/FROM NAMED hides it away from
the set of named graphs. Although it can be referenced explicitly in
FROM NAMED to access it as a named graph.