tests: Drop test poking fts:snippet() limits
This is actually handled internally as an error, but never propagated.
This will change in later commits, so drop this last test expecting to
handle fts:snippet() calls with a non-positive number of tokens in a
meaningful manner.

Later commits will also poke for this turning into a hard error.