ci: Prefer libsoup from submodules in Debian jobs
The libsoup version currently present in Debian Bookworm does
exhibit some bug preventing our test suite to run cleanly.

This seems to be a minor difference irrelevant to us (extra
newline at the end of HTTP responses), so don't bump our
version dependency, but uninstall libsoup in these jobs so
that meson prefers building it as a subproject.

While at it, remove some cruft that was added in our last
stages of trying Ubuntu for testing, and ensure that gcovr
does not get coverage info from subprojects.