tests: Test more paths in SERVICE {} handling
In some paths with non-commutative operations like optional/minus,
we have got to put a SERVICE{} query in the first place, resulting
in bindings handled as strings, and resources being matched and
converted to strings within the query. Test that this works.

Also test multiple variables in a SERVICE{} clause, constraints,
and different types.