http: Fix double free, and handle cancellation better
If we obtain an error when reading the content for a HTTP response
(a possible scenario is cancellation), we try to free the
TrackerHttpRequest struct twice.

Handle cancellation more conventionally, which means the data
will already be freed beforehand in that case, and avoid the double
free on any (other) error.

Fixes: ac8a35747 ("http: Send HTTP response data using chunked encoding")