core: Handle errors in parsing content of collections
The handling of collections needs to allocate some memory that needs
to be freed in error paths. This means we will leak by using the helper
define, as it return in place and will not let us handle this leaked
memory. Avoid the helper and use the function directly so we can handle
the error paths.