library: Fix official Turtle extension for ontology files
Since 3.8.x we allow ontology files to have other extensions than
".ontology" and RDF formats than Turtle.

But as a bit of a brain fart, we've gone with the ".turtle" file
extension for the Turtle format. The official file extension for
this format is actually ".ttl". Change it in documentation and
code, while this blunder still has got short legs.