tests: Move SPARQL test ontologies to a different folder than test data
In some places, we mix .ontology files used to build the test database
with .ttl files that are loaded later on during the test run. This may
lead to duplicate data and broken test expectations if the library starts
considering .ttl files as suitable ontology files (thus implicitly loaded
during database creation).

Move the ontologies for these tests to a separate folder, so that the
distinction is clearer between ontology and bulk data to be loaded
afterwards.