fuzzing: Add targets for oss-fuzz to run
Now that TinySPARQL is included in oss-fuzz, add some targets for it
to run. This support is heavily based on glib infrastructure, which
besides optionally allowing them to run under proper
llvm+sanitizer+fuzzingengine environment, it runs the tests with
a fixed random file to ensure they won't crash out of the box.

So far, 4 tests were added:
- One for SPARQL queries and updates
- One for RDF data in Turtle format
- Same for Trig
- Same for JSON-LD

A possible future additional test is database creation from ontology
definitions, although we don't have the necessary API at the moment
to test this from in-memory input.