common: Use broader gio dependency in parser modules
In builds using glib as a subproject, these modules indirectly rely
on headers pulling gio.h, which indirectly pulls gmodule.h. Since in
the subproject gmodule.h will not be in the same place as glib.h, we
will fail to see it since we do not explicitly link to gmodule.

Use the more honest gio dependency, which will pull gobject and gmodule
dependencies and add the necessary -Isubprojects/glib/gmodule/ path.