build: Add builtin_modules build option
This option will make our existing modules be compiled directly
into the library. It will be useful in environments where fully
static builds are necessary.

Currently only works for parser modules, leaving http modules aside,
but that's all we need at the moment for oss-fuzz to fuzz. More
global support for this toggle is a nicety that will come in the
future.