fuzzing: Disable SQLite lookaside memory allocator
This is a per DB connection cache to avoid too frequent allocations.
OSS Fuzz does not seem to like it, on one hand complains that there
is a "global state" surviving between runs (which should not be
true...?), and on the other hand it may lead to OOMs given the OSS
fuzz memory limits if there are many concurrent fuzzing operations
going on at the same time. Disable this lookaside table so memory
does not balloon up as quickly.

While at it, also close the connections explicitly, so that it's
not up to refcount being fully correct.