library: Register GResources manually
Automatic registration of GResources does not work on all compilers
and environments. A notable one is building with -Ddefault_library=static
as that will make the final executables lack the expected ontology
resources.

We want to support static compilation for OSS Fuzz, make registration
manual so that this data is not snipped from the final executable.