core: Fix error handling reading SPARQL update string
The error handling paths in this function will only return FALSE
if the inner error is set, we were however bypassing it in this
check.

This could make the function spuriously return TRUE, even though
the error was set, causing the error to be silently ignored, or
possibly causing further runtime warnings while finding future
errors (e.g. closed stream while handling a batch of updates)
and finding the GError was already filled in.