fuzzing: Reduce startup time for fuzzing tests
It does not matter as much what ontology does the database support,
so perhaps try not loading an ontology with hundreds of tables. This
should improve startup time.

Also, make the connection static so it can be reused in multiple runs,
which should favor how oss-fuzz runs these tests in practice.