core: Avoid quadratic behavior in parser extension
The fn:string-join xpath function is a funny one, per spec it does
take the strings to join as sequence surrounded by its own set of
parentheses, e.g.:

  SELECT (fn:string-join (("crazy", "ain't it"), ",") AS ?what) {}

This extra level of argument lists is outside the SPARQL parser specs,
but we've supported this since very early times. When the parser was
refurbished during 2.x, we opted to handle it as an ArgList inside
an ArgList, as a Tracker extension.

This however may incur in unintentially quadratic behavior handling
this arglist for a single argument case, since this matches both
BrackettedExpression and the nested ArgList. Since both contain have
a fairly deep decision tree, and can contain further functions with
further arguments, each level of depth doubles the amount of checks
to do.

Nowadays, fn:string-join is fringe functionality with the CONCAT()
builtin SPARQL call, and we do certainly not need to support full
expressions as arguments, but can compromise to just supporting
RDF literals and variables to avoid this potentially deep side
choice in the parser tree.

oss-fuzz#395921922