fuzzing: Clear database on successful batches
If OSS fuzz is trying to tap on permutations of things that turn out
successful after all, there is some room for this data to accumulate
until the test "runs out of memory" in OSS fuzz infrastructure. Let
the database be cleared in these situations, so any successful batch
triggers a clean database for the next run.

Hopefully this will not reintroduce "timeouts" due to the database
requiring to be created from scratch on virtually each run...