core: Add error limit on parser
Add an error counter, so that it will be bumped each time the
parser gets stumped in the same point of the SPARQL query again,
and reset when it manages to advance. And add a (very permissive)
limit on errors, so we bail out before attempting every possible
path with soemthing that is in all likelyhood incomplete or invalid
SPARQL at that point.

This generically fixes a small class of bugs with some of our
pre-1.1 SPARQL extensions inadvertently exposing quadratic parser
behavior when testing all TrackerGrammarNode paths.