fuzzing: Set size limits in fuzzing targets
We are currently at an arms race with OSS fuzz cranking up the
sample size to trigger "timeouts" with some of our fuzzing targets.
These timeouts also must happen under heavy instrumentation and
busy machines, since testing the same samples locally do result
in a delay measurable with a wall clock, but still in the range
of tens of seconds instead of thousands.

The library has virtually no upper bounds in the data managed, and
some operations will just be inherently slow. OSS fuzz is not here
to test that, and its documentation acknowledges that:

https://google.github.io/oss-fuzz/advanced-topics/bug-fixing-guidance#should-all-reported-issues-be-solved

So set up a generous limit of 800KiB for all fuzzing targets, any
pathologic behavior (or crash, etc) should be spotted with way less
than that.