core: Reduce lifetime of resource IRI->ID cache
With large RDF data, it is unlikely that we need access to each and
every resource referenced within the transaction, and even with smaller
RDF data, it is unlikely to reference the same IRIs over and over.

Clear the cache every time we flush into the WAL log, so that it's
cleared frequently and the in-memory cache does not possibly explode
with large documents and/or IRIs.

oss-fuzz#405187821