core: Use helper function to update single-valued properties
Use the new helper function underneath, which will perform the checks
based on the old/new value.

The idea with this helper is that we will leave all consistency checks
to "the database engine", so this also means we cannot coalesce operations
on a same single-valued property at the code level. These don't happen
often, so there shouldn't be much of a performance hit due to splitting
operations on a same table while dealing with correct SPARQL that will
succeed insertion.