core: Avoid reading values for property updates
We have now a mechanism to handle cardinality errors and other
possible mishaps when dealing with single-valued properties, and
the SQL queries dealing with multi-valued properties are already
prepared for not requiring to get previous values.

This leaves other scarce situations where we do read existing
property values upfront: When dealing with refcounting and when
dealing with propagation of domain indexes. For this we must still
feed in the inserted values to the cache.