core: Avoid getting old values for refcount handling
Let the database update the relevant rowids, without querying
old values through code, while still coalescing the operation
locally in the batched changes.

Since these refcounting changes are queued up as operations
in the log, we will be executing exactly one SQL update for each,
this also means we do now rely on the refcount trigger to handle
deletion of rows where refcount dropped to 0, as we cannot issue
a second DELETE query to remove the row from here.