core: Create all graphs within the same database file
So far, we have implemented graphs as independent databases holding
the same structure as the main database (as defined by the ontology).
This has multiple limitations, softer in SQLite as a build-defined
limit on the amount of opened databases (10) or harder as an OS-imposed
limit on the amount of file descriptors a process can open.

Switch (in one go, since it's hard to do incrementally without intermedate
brokenness) to creating these graphs as sets of tables with a common prefix
in the same database file.

This has some great pros:
- Makes TinySPARQL databases contained in a single file
- Massively increments the natural maximum amount of graphs, since we just
  need a bunch of extra pages for a graph, not a new file and scarce
  FD slot.

And only one smaller con:
- With the multi-file approach, and insane database limits in SQLite we
  could store up to 320 TB of data (32TB per database, multiplied by 10
  graph databases). This gets reduced to the more "modest" maximum size
  of 32TB collectively for all graphs. This should still be enough for
  the usecases we intend to cover with TinySPARQL, so hopefully not a
  big drawback.

Other limits still apply (e.g. querying multiple graphs still involves
poking multiple tables, now in the same database file instead of different
ones).

This commit:
- Changes all the places necessary to access per-graph tables through the
  new prefix instead of through database files.
- Bumps the database version
- Adds migration code so we can convert databases from the previous
  format(s)