library: Do not let @base base upon itself in Turtle documents
This is not explicitly defined to be the behavior in the Turtle specs,
the section https://www.w3.org/TR/turtle/#sec-iri-references defines
the URL to be relative according to the
document/encapsulation/retrieval/default layers (RFC3986 Section 5,
"Reference resolution", which we don't observe), but not relative to
a prior existing base. Implementation-wise, we are also apparently left
alone, with other Turtle implementations I tried replacing the existing
base altogether.

Switch to replacing the base, instead of accumulating over it. The
resulting IRIs will be checked to be correctly constructed absolute
ones in lower layers. As we don't need string accumulation, we can
switch the storage format to a char* instead of a GString.

We had a test poking at this misbehavior, also change it so it keeps
testing for @base changes without relying on relative IRIs to be a
valid base.

oss-fuzz#405187821