core: Simplify insert/update/commit/rollback callbacks
It is unthinkable that a subscriber of events might want these
individually, merge insert/delete and commit/rollback into two
sets of callbacks, that are called with an enum value.

Also update the hook points in the direct connection implementation,
as that's the only direct user of this API in core, in order
to implement TrackerNotifier.