core: Account for additional field in propagation queries' hash/equal functions
Generically, we compare the first field of a TrackerDataLogEntry. The addition
of new operation types missed that this will not be enough for all the propagation
queries that do depend on two fields to produce the update statement, namely
when propagating domain index values, and multi-valued property propagation.

This will fix statement MRU lookups maybe picking the wrong statement for some
operations, e.g. happening on the wrong table.

Fixes: 281423cd9 ("core: Avoid getting properties to propagate domain indexes")
Fixes: b00b1c943 ("core: Log propagation of property values")