tests: Add tests for database refcounting
An indirect effect of refcount being correct is that URNs of no
longer existing resources will be trimmed on database close. We
can indirectly test that this happened through "tracker:id(<urn>)".

Add some tests adding and removing rdfs:Resources in various
simple scenarios, in order to ensure resources are trimmed as
expected.