library: Allow the store location in local constructors to define a filename
Since we used to store graphs as multiple databases, we declared this location
to be a directory, where the default "meta.db" and additional graph database
files were created.

Now we store everything in a single file, and the misc files can be built
from it (SQLite does itself, with shm files). We can allow this store location
to describe a database filename, for cases where multiple databases might
be desirable in a single directory.

We now check for an extension being present, and respect the application
choice in that case. For backwards compatibility, we still favor pre-existing
directories.

This also requires changes in the way we handle our ancillary files, mainly
for tagging databases as corrupted, this is now done with a file that is
relative to the database file basename.