core: Give an specific "application ID" to TinySPARQL databases
Now that we store databases altogether in a single file, we are (arguably)
a file format. Provide a distinct application ID ('Spql' encoded as a fourcc
number) to our database file, so it will eventually become possible for
`file` and other identification tools to distinguish TinySPARQL files
from other SQLite databases.

While there does not seem to be any install location for this kind of files,
also add a utils/magic.txt file containing the magic(5) format necessary
for file (the CLI utility) to identify TinySPARQL databases specifically.