core: Add heuristic to bail out earlier on unparseable queries
Add a heuristic to avoid expensive retries on eminently unparseable
queries, it goes like this: It is a natural thing that we have to
rewind and retry different alternatives, it's however an indication
of hard failure if we have to rewind a massive portion of the parser
state to retry parsing a big portion of the query. This heuristic
limit is set at half a million parser nodes, which is also an indication
of someone trying to exploit pathologic cases, like it's the case of
OSS Fuzz.

oss-fuzz#402939128