core: Avoid pointless ref/unref pair on TrackerDBStatement
We used to keep a ref and flag a TrackerDBStatement with stmt_is_used
while the statement has been executed and is being iterated. We do not
need this ref, and already assert that the statement is not in use
when it is being finalized (which would indicate misuse).

We do not need this ref/unref pair, and given correct behavior the
statement will be naturally kept alive. This ref/unref pair also seems
to lead Coverity into thinking the TrackerDBStatement might be freed
early, which is not the case.

CIDs: #1643031, #1643032