libtinysparql: Fix thinko in TrackerNotifier
Use the n_arg as the baseline for statement argument numbers,
not the current "first" position in the array (i.e. the first
position that should be inspected in the next query to extract
tracker:uri).

This fixes warnings like:

Tracker-CRITICAL **: 23:08:57.748: Could not get cursor: Parameter 'arg1' has no given value