core: Pass ontology as a TrackerDeserializer
For the existing API, this will typically be a
TrackerDeserializerDirectory with the user-provided location,
merged with another such deserializer for the base ontology.

However, we can only afford importing RDF data with file based
streams, since any other deserializer (e.g. pipe based) cannot
be rewound for this second pass. This is, so far, always the
case.