core: Fix possible confusion handling refcount of multi-valued properties
Confusingly, the multi-valued tables have a ROWID column (as usual, unique
and primary key), and an ID column (non-unique, refers to the resource having
a multi-valued property value). SQLite normally allows usage of ROWID and ID
interchangeably, but this distinction is crucial when dealing with these
tables refering to multi-valued properties.

This might result at worst in the wrong resource IDs having their reference
count decreased, if the ID-as-ROWID matches something in the table. At best,
it would result in a reference leak on the expected resource.