core: Fix refcount with idempotent property modifications
When deleting a property, we were using operations that cleared all
values even though we were deleting a single specific property value.
Add specific refcounting operations for property values that may or
may not be there, and use it to modify the refcount only if the
property deletion will have an effect.

For property insertions, we are inconditionally adding an extra
reference at the end of the operation, start it with another such
refcounting operation, in order to ensure idempotent inserts will
not result in an increased reference.